Remove default static route on vrouter and centralpoint nodes used in OpenStack environments
OpenStack has the ability to add extra static routes to subnets.
Unfortunately it is not possible to add those routes only to specific
ports, so the OpenVPN client cannot set its own route.

This way, OpenStack sets the route for all nodes, and during the vrouter
instantiation this route is deleted before starting up the OpenVPN
client.